LIMA-4: Big lima_search update
Refactored search_dir() to be recursive (to support the new -r argument)
Discovered a UTF-8 BUG(?) in search_file() once it started recursing
directories.  search_file() raises a UnicodeDecodeError exception when
attempting to read a non-UTF-8 encoded file (go figure).  For the time
being, I circumvented this very specific exception with
straight-to-stderr error message and added LIMA-5 to the backlog.

Everything appears to be working just fine.

Changes to be committed:
	modified:   lima/lima_search.py

Changes not staged for commit:
	modified:   lima/lima_args.py
	modified:   lima/lima_main.py