LIMA-5: Fixed(?) BUG with search module refactor
GOOD NEWS: UnicodeDecodeErrors now get shunted to a byte reader that
doesn't care so much about UTF-8.

BAD NEWS: Any case-insensitivity is lost when transitioning to reading
bytes.  Any case-sensitive byte-reading functionality is "not worth the
squeeze" (if it's even reasonably possible).

This is, to quote my father, "Good enough for government work."

Changes to be committed:
	modified:   lima/lima_search.py